requirements.txt
nemo_toolkit
pytorch_lightning
transformers==4.40.0
torch 
torchvision 
torchaudio
hydra-core
transformers
sentencepiece
soundfile 
librosa
pandas
huggingface-hub==0.23.2
youtokentome
braceexpand
webdataset
h5py
psutil
ijson
sacremoses
matplotlib
ffmpeg
Cython 
packaging
sacrebleu
rouge_score
einops
jieba
opencc
pangu